+++ src/pysight/allocation_tools.py
@@ -34,6 +34,7 @@ class Allocate(object):
     censor             = attr.ib(default=False, validator=instance_of(bool))
     dict_of_data       = attr.ib(default={}, validator=instance_of(dict))
     tag_interp_ok      = attr.ib(default=False, validator=instance_of(bool))
+    tag_to_phase       = attr.ib(default=True, validator=instance_of(bool))
     sorted_indices     = attr.ib(init=False)
 
     def run(self):
@@ -96,7 +97,8 @@ class Allocate(object):
         else:
             print('Interpolating TAG lens data...')
             tag_pipe = TagPipeline(photons=self.df_photons, tag_pulses=tag, freq=self.tag_freq,
-                                   binwidth=self.binwidth, num_of_pulses=self.tag_pulses)
+                                   binwidth=self.binwidth, num_of_pulses=self.tag_pulses,
+                                   to_phase=self.tag_to_phase)
             tag_pipe.run()
             self.df_photons = tag_pipe.photons
             self.tag_interp_ok = tag_pipe.finished_pipe